requirements.txt
flask>=3.0.0
flask-login>=0.6.0
python-dotenv>=1.0.0
qrcode[pil]>=7.0.0
pillow>=10.0.0
openpyxl>=3.1.0
plyer>=2.1.0
